servlet 3.0 support. WICKET-3147
git-svn-id: https://svn.apache.org/repos/asf/wicket/trunk@1031086 13f79535-47bb-0310-9956-ffa450edef68

+++ wicket/src/test/java/org/apache/wicket/protocol/http/WebResponseExceptionsTest.java
@@ -36,10 +36,12 @@ public class WebResponseExceptionsTest extends WicketTestCase
 	 */
 	public void testBufferedExceptionErrorPage()
 	{
-		tester.getApplication().getRequestCycleSettings().setRenderStrategy(
-			IRequestCycleSettings.RenderStrategy.REDIRECT_TO_BUFFER);
-		tester.getApplication().getExceptionSettings().setUnexpectedExceptionDisplay(
-			IExceptionSettings.SHOW_EXCEPTION_PAGE);
+		tester.getApplication()
+			.getRequestCycleSettings()
+			.setRenderStrategy(IRequestCycleSettings.RenderStrategy.REDIRECT_TO_BUFFER);
+		tester.getApplication()
+			.getExceptionSettings()
+			.setUnexpectedExceptionDisplay(IExceptionSettings.SHOW_EXCEPTION_PAGE);
 		testInternalErrorPage();
 	}
 
@@ -48,8 +50,9 @@ public class WebResponseExceptionsTest extends WicketTestCase
 	 */
 	public void testExceptionErrorPage()
 	{
-		tester.getApplication().getExceptionSettings().setUnexpectedExceptionDisplay(
-			IExceptionSettings.SHOW_EXCEPTION_PAGE);
+		tester.getApplication()
+			.getExceptionSettings()
+			.setUnexpectedExceptionDisplay(IExceptionSettings.SHOW_EXCEPTION_PAGE);
 		testInternalErrorPage();
 	}
 
@@ -84,6 +87,6 @@ public class WebResponseExceptionsTest extends WicketTestCase
 		AjaxLink link = (AjaxLink)tester.getComponentFromLastRenderedPage("link");
 
 		tester.executeAjaxEvent(link, "onclick");
-		assertEquals((Integer)500, tester.getLastResponse().getStatus());
+		assertEquals(500, tester.getLastResponse().getStatus());
 	}
 }
