+++ wicket/src/test/java/org/apache/wicket/redirect/abort/AbortExceptionTest.java
@@ -38,7 +38,7 @@ public class AbortExceptionTest extends WicketTestCase
 	public void testNoAbort()
 	{
 		tester.startPage(AbortExceptionPage.class, new PageParameters().set("trigger", false));
-		Assert.assertEquals((Integer)1234, tester.getLastResponse().getStatus());
+		Assert.assertEquals(1234, tester.getLastResponse().getStatus());
 	}
 
 	/**
@@ -49,7 +49,7 @@ public class AbortExceptionTest extends WicketTestCase
 		try
 		{
 			tester.startPage(AbortExceptionPage.class, new PageParameters().set("trigger", true));
-			Assert.assertEquals((Integer)1234, tester.getLastResponse().getStatus()); // this will
+			Assert.assertEquals(1234, tester.getLastResponse().getStatus()); // this will
 			// fail
 		}
 		catch (RuntimeException x)
